remove byteorder dependence
